requirements.txt
# Django основные пакеты
Django==5.0.1
djangorestframework==3.14.0

# База данных
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Работа с файлами
python-docx==0.8.11
openpyxl==3.1.2
xlsxwriter==3.1.9

# Фильтрация и пагинация
django-filter==23.3
django-crispy-forms==2.1

# Валидация и формы
django-bootstrap5==23.3

# Аутентификация и авторизация
django-allauth==0.57.0

# Загрузка файлов
Pillow==10.1.0

# Разработка и отладка
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Производство
gunicorn==21.2.0
whitenoise==6.6.0

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Линтинг и форматирование
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Проверка безопасности
safety==2.3.5
bandit==1.7.5

# Прочие утилиты
python-decouple==3.8
django-cors-headers==4.3.1
django-cleanup==8.0.0 